requirements.txt
Flask>=2.0
Werkzeug>=2.0
mysql-connector-python

Flask
mysql-connector-python
python-dotenv
